screen/requirements.txt
Pillow >= 9.3.0
notion-client >= 2.0.0
python-dotenv >= 0.21.0